Organize GridDividerDecoration into util.ui package.
