+++ app/src/main/java/com/bsstokes/bsdiy/ui/GridDividerDecoration.java
@@ -1,4 +1,4 @@
-package com.bsstokes.bsdiy;
+package com.bsstokes.bsdiy.ui;
 
 import android.content.Context;
 import android.content.res.TypedArray;
@@ -7,6 +7,8 @@ import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import com.bsstokes.bsdiy.R;
 
 /**
  * ItemDecoration implementation that applies and inset margin
